GUI: Some fixes and multi line text node rendering

+++ src/tdme/gui/nodes/GUIMultilineTextNode.h
@@ -55,7 +55,7 @@ private:
 	bool parentOffsetsChanged;
 	float parentXOffsetLast;
 	float parentYOffsetLast;
-	float yLast;
+	int yLast;
 	int charStartIdx;
 	int charEndIdx;
 	int widthLast;
